requirements.txt
Pillow>=10.0.0
img2pdf>=0.5.0
Flask>=3.0.0
